Decode Android manifest using Jackson
ManifestData (as returned by AndroidManifestParser.parse()) is missing
some information we need to improve parcelo. Namely, it doesn't include
non-custom requested permissions, which we'll need later. Additionally,
it exposes no access to split names, making it more difficult to
determine whether the current APK is a base APK or a split APK.

Using a custom data class will allow us to add these attributes (and
others, such as usesCleartextTraffic) as we need them without
restriction. We'll also be using libraries actually intended to be
consumed by us, which is a nice bonus.