Trust all javadoc and sources artifacts
There doesn't seem to be a way to make Gradle generate checksums for
these automatically, and IDEs like to automatically download javadoc and
sources artifacts upon Gradle sync, so IDE users run into dependency
verification issues often. There isn't really a reason to verify the
javadoc and sources checksums, so it's safe to trust them for the
benefit of IDE users.