Improve database consistency of app updates
A reviewer shouldn't be assigned to an update without a review issue
group, and likewise, a review issue group shouldn't be assigned to an
update without a reviewer. This is already enforced in the application
logic, but for better security and robustness it's a good idea to
enforce it at the database level as well.