Fix permission review bypass
Because of
https://github.com/FasterXML/jackson-dataformat-xml/issues/275, declared
permissions which are interleaved in other fields are not treated as the
same list. Critically, Jackson's behavior here is to overwrite the
previous list items when it encounters a "new" list of <uses-permission>
items, which is discrepant from the Android package parser.

The result of this behavior is that apps which declare their manifest
permissions in a non-contiguous list (which the standard Android tools
sometimes causes, although a malicious actor could certainly modify the
binary Android manifest to trigger this behavior at will) will not have
all of their permissions scanned by the review system.

To work around Jackson's behavior, we create a custom setter as per
https://github.com/FasterXML/jackson-dataformat-xml/issues/275#issuecomment-352700025
which appends to the permission list rather than overwrites it. This is
admittedly a bit of a hack, but we're not setting this field manually
anyway and there doesn't seem to be another solution at the moment
besides using a different XML parsing library.