Set session cookie lifetime to max in development mode
Session expiry times are set to a maximum value in the database for
development mode as of "04f8d3c: Set debug session lifetime to
(practically) infinity", but SESSION_LIFETIME is still used as the
cookie expiry time, rendering the change ineffective for clients which
respect the cookie expiry time. This change ensures the expiry time of
the cookie is just as consistently high in development mode.