Validate parsed app IDs
Although the standard Android build tools ensure the user-specified app
ID is valid on build, it is possible to modify it afterward using a hex
editor or similar to an invalid value. This could be abused to trick the
developer console into accepting an app ID which won't be accepted by
client OSs.

Furthermore, since we use parsed app IDs in URL path construction, this
behavior is a potential vector for path traversal.

Validate the well-formedness of app IDs to ensure that 1) clients will
be able to install the app as expected and 2) we can safely use app IDs
in URL construction.