Create newtype wrapper for AppId
Android app IDs are strings which follow a specific well-defined format.
Currently validation on supposed app ID strings is performed manually
before the strings are used. This approach leaves the risk of
accidentally using a string as an app ID before validating it, allowing
arbitrary data in place of well-formed data and so violating assumptions
made about the structure of the string, potentially leading to a
vulnerability.

This change refactors parcelo to use a newtype, AppId, which represents
a well-formed Android app ID. The only way to construct this type is
with the AppId.parseFromString function, ensuring validation occurs
before use of a string as an app ID.