Properly mark no-review-required updates as submitted once submitted
Updates are currently not marked as submitted once submitted if they
don't require review since they begin publishing immediately. However,
now that we allow viewing an update's status in the API, the API
incorrectly displays updates in the process of being published as
unsubmitted if they didn't require review. Fix this by marking the
update as submitted regardless of whether it requires review.