Fix failed draft deletion
When deleting a draft, we accidentally were looking up the icon file ID
to delete with the ID of the icon itself. This happens to work in a
couple of test cases, but breaks once a draft has been deleted, another
added, and another draft deletion is attempted. Properly delete the icon
file and thus the draft.