Fix database lock conflicts
This change fixes existing database lock conflicts by making two
different but related changes:

- prevent nested transactions by moving JobRunr job creations outside of
  Exposed transactions
- modify app update locking mechanism to be more robust

The first change is self-explanatory. The second involves adding an
"updating" field to the App entity which is modified whenever an app
update publishing job is registered for the corresponding app or when
such a job completes. This field is checked whenever an app is submitted
to be published without review or when an approving review is created.
This check prevents multiple update jobs for a given app from existing
at once and contending with each other while the submitter or reviewer
incorrectly sees that submission/review succeeded.