Fix fatal TLS handshake exception when publishing
The Ktor HTTP client throws a fatal TLSException HandshakeFailure when
connecting to the repository. Curiously this doesn't occur in our
staging environment, but can easily be reproduced locally.

The issue seems to find root in the CIO Ktor client's cipher support
suite as indicated by issues with the same error such as [1]. Supposedly
this issue was resolved, but others such as [2] indicate the issue is
still present despite it being closed upstream. Without useful error
information from the stack trace, the best option seems to be switching
to another well-supported client engine such as Apache5. As a bonus,
Apache5 also has HTTP/2 enabled by default.

[1]: https://github.com/ktorio/ktor/issues/439
[2]: https://github.com/oliver-charlesworth/craft-watch/pull/198