Add short description edit creation and viewing support
This change introduces a new concept called an edit. An edit is tied to
a published app and contains a set of requested changes to app metadata
which occur atomically when applied. This way, e.g. app icons and
descriptions can be updated simultaneously. Edits require a reviewer
approval, after which they will be published. For now, we introduce
edits with only a single field - an app's short description - but we
will add more fields later.

App updates should eventually be folded into edits so that app updates
can occur simultaneously with app metadata updates. However, this is not
critically important and so is considered out of scope for this
changeset.