Fix console crash on second run with on-disk database
At the beginning of a debug run of the developer console, we create a
user and reviewer for local testing purposes. We've used the
insertIgnore function when creating these so that when persisting state
to an on-disk database, the second run doesn't crash due to attempting
to insert a duplicate database entry. However, when an entity already
exists, insertIgnore returns the entity it *would have* created, not the
preexisting entity. Thus, a new reviewer was created on each subsequent
run of the developer console in debug mode since the user ID it was
referencing was always unique.

This is a foreign key constraint violation since the user IDs each new
reviewer was referencing didn't actually exist. However, this bug didn't
cause a crash until we enabled foreign key constraint enforcement in
"d965c2b: Enforce SQLite foreign key constraints", so it went unnoticed
until now.