Build console KDoc in CI
Because there is little console KDoc at the moment, all of the console's
packages are excluded from requiring docs. The intent is that the
exclusion regex is updated as individual sub-packages are documented
until it can be removed entirely.

We could require docs for the repository server as well, but this
doesn't make sense for now since we may be phasing out the repository
server in the near future.