Unconditionally mark cookies secure
Ktor 2.3.8 and earlier didn't provide support for marking cookies secure
unless the transport was known to be secure by the framework, either
through enabling TLS in the application itself (which we don't do since
we terminate TLS at a reverse proxy), or by enabling X Forwarded headers
support and passing the X-Forwarded-SSL header in the reverse proxy.

Thankfully Ktor 2.3.9 removes this restriction, allowing us to enable
secure cookies over HTTP, allowing for a minor simplification of the
relevant code.