Fix incorrect publishing of x86_64 ABI splits
Because the world is cruel, Android's supported ABIs as reported by the
Android platform are as follows:

- armeabi-v7a
- arm64-v8a
- x86
- x86_64

Notably, the split names of each of these ABIs as reported by their
corresponding split APKs are "armeabi_v7a", "arm64_v8a", "x86", and
"x86_64". However, their names as reported by the Android platform use
hyphens instead of underscores... except for x86_64. Thus, we thought we
could be clever and replace the underscores with hyphens to match the
split names with the values on the Android platform, unknowingly
breaking ABI split publishing _only for x86_64_; and woe is me, trying
to test downloads on an x86_64 emulator.

For the record, I think this is at least the third time I've fixed
hypen/underscore confusion breaking downloads.