Revamp APK set parsing
The current APK set parser is convoluted and not easily modifiable to
support feature modules, archived APKs, and other features of APK sets.
Nearly all information about the APK set is parsed from the APKs
contained therein, and various incorrect assumptions are made about the
APK set's contents.

This new parser takes a different approach. All data which can safely be
trusted from the APK set's table of contents is used, while only data
which must be taken from the APKs (either because it's not present in
the table of contents or must be taken from the APKs for security
reasons) is. In this way, the parser remains safe against malicious
inputs (e.g. a table of contents with a different app ID than the
associated APKs, which could result in the console overwriting another
app when publishing) while still taking advantage of the table of
contents, resulting in a much simpler parser.

Developers are still able to mess up their own app's publication, but
only if they make modifications to their APK set outside of bundletool.
We may make improvements to prevent even this in the future.

Because of how this parser works, APK sets can no longer be parsed in
memory and are instead parsed on disk through a temporary file. We
wanted to do this eventually anyway because of the memory savings.

Bundletool version checking was removed because it served no particular
purpose and can be trivially bypassed. We may reintroduce it in the
future to give more helpful error messages to developers using
bundletool versions unsupported by Parcelo.

Closes #229