Fix icon parts not being properly disposed
Multipart form files, when read via `readAllParts()`, are saved to disk
in a temporary location and must be manually disposed. Unfortunately,
Parcelo fails to dispose of app icons and additionally fails to dispose
of APK sets if an exception is thrown while parsing them.

Ensure multipart form files are disposed of properly in these cases.

Note that if `readAllParts()` itself throws an exception, files may not
be disposed of properly. We should probably migrate to `forEachPart {}`
in the future.