Use db as source of truth for repository metadata
The current behavior of S3PublishService when publishing an edit is to
fetch the old repository metadata from the S3 bucket, patch it with new
data, then republish the patch metadata over the old metadata. The
problem with this approach is that it requires unnecessary data fetches
and doesn't maintain a well-defined source of truth for the repository.

Instead, store generated repository metadata in the database and patch
_that_ when publishing edits so that we can consistently source the
repository from a single location. The data is relatively small, so
storing it in the database shouldn't have a significant performance
impact. If we really need to in the future, we can store the metadata as
files in a separate storage bucket.

For consistency, additionally update the database repository metadata
when publishing app updates.