Move private storage to S3-backed implementation
As part of our containerization efforts, it is necessary for us to not
rely on the local filesystem for persistent storage. S3 is the standard
vendor-agnostic protocol for this, and there are multiple open-source
self-hostable implementations of S3 servers besides.

Unlike ce76e9d: "Use db as source of truth for repository metadata",
this change doesn't contain an automatic pseudo-migration which, in this
case, would move files from local storage to a private S3 bucket. This
is because we ran into opaque errors relating to SQLite locking and
would rather implement this data migration manually via a script. As a
result, the release containing this change will require manual
intervention to migrate all existing data.

As part of making this change, multiple issues related to file storage
and deletion were discovered and fixed. The first is that the job
functions `cleanFile()`and `cleanDeletedFiles()` were tightly coupled to
the local storage implementation. We resolved this issue by making them
generic over the file storage implementation. The second is that draft
components weren't deleted in the proper order, causing foreign key
constraint violations whenever the job scheduler attempted to delete
zombie file entries, explaining the logs we've been seeing in
production. Drafts are now deleted properly. Notably, icon entries
weren't deleted at all, so these are now deleted as well.

Overall, this change should result in much more robust handling of
private file storage and fixes a handful of existing issues relating to
file management in general.