Fix foreign key constraint violation when publishing updates
When publishing an update, the corresponding app's file ID is set to
point to the file of the new update, and the old file is deleted. This
works because no other entity references the app's old file; the draft
which originally referenced it is deleted. However, once a _second_
update is published for a given app, attempting to delete the app's old
file results in a foreign key constraint violation because it is
referenced by the previous update.

As a temporary solution to this, allow the file ID of updates to be
null and set it to null when the underlying file is deleted. Ideally
we'll handle this in a cleaner way in the future, but this approach
solves the immediate issue in a way that's not intrusive.

This issue was found when testing dumping/import from SQLite to
PostgreSQL, as SQLite didn't catch this constraint violation even with
foreign key constraint enforcement enabled. It appears that PostgreSQL's
data integrity capabilities are already paying off even if the original
issue was a case of PEBCAK.