Ensure S3 storage service only cleans files marked deleted
There was a bug in the initial implementation of the S3 object storage
service in 65c5f29: "Move private storage to S3-backed implementation"
which allows `cleanFile()` to clean files which aren't marked as
deleted. This isn't a significant issue since we don't currently rely on
it only functioning on files marked deleted, but the latter behavior is
in the API documentation, consistent with how the sister function
`cleanAllFiles()` behaves, and helpful for preventing accidental file
deletions, so we should fix it.