Fix bug causing app parsing to always fail
It seems that a recent Kotlin upgrade between 2.0.20 and 2.2.0 has made
breaking changes to annotations, causing breakage in apksparser.

The `@JacksonXmlProperty` annotation on `AndroidManifest`'s usesSdk
field was applying only to the value parameter of the constructor and no
longer applying to the field itself. Jackson observes the annotation
only when it applies to the field, so it was ignoring it, making Jackson
search for an XML tag named "usesSdk" (which doesn't exist) instead of
"uses-sdk". This resulted in target SDKs not being parsed, and since
they're a required attribute for app parsing to succeed, app parsing was
failing, completely preventing app or app update uploads.

The fix is to explicitly tell Kotlin we want this annotation to apply to
the field instead of the value parameter.