Raise maximum file upload size to acceptable level
Ktor v3 introduced a new default 50 MB limit for binary and file uploads
in multipart data. Uploaded APK sets are very commonly much larger than
this, so we need to bump the size to an acceptable number to be able to
accept apps.

See https://ktor.io/docs/migrating-3.html#new-default-limit-for-binary-and-file-items
for info on the new limit.

1 GiB was chosen since it should give a safe buffer above the apps we've
been processing lately. See accrescent/meta#33 for progress on setting
more concrete app size limits.