Publish apps to directory service instead of repository
This change makes Parcelo publish new apps, app updates, and app listing
edits to the directory service instead of the repository. It is
completely backward-compatible with data from existing installations of
Parcelo, but it is _not_ backward-compatible with existing publication
methods, i.e., with this change, Parcelo no longer publishes updated app
information to the now-deprecated repository.

This change is fragile and not the highest quality. It is intended to
exist only temporarily while Parcelo receives a much larger revamp. It
may cause Parcelo crashes under adverse conditions, and there is some
extensive code duplication in the directory service publishing code.

Other noteworthy changes included in this change include the following:

- app update APK sets are no longer deleted from disk once published
- published APKs are no longer deleted from public object storage once a
  new version of the corresponding app is published
- some imports were moved due to IDE formatting

We need to keep these old files around to ensure that external file
references can't be orphaned and that clients can access stale APKs when
relying on caching.

Closes #714