Remove directory service migration code
All apps have been migrated to publish via the directory service, so
there is no need for this code anymore.

We can't delete migrateAppsToDirectoryService() yet since it is called
reflectively by JobRunr, so this change attempts to delete the recurring
job on startup so we can remove the function later. We can, however,
remove the function's implementation now since it is no longer useful.