Set default transaction isolation level to SERIALIZABLE
SERIALIZABLE is a much safer and more consistent default than the
typical READ COMMITTED, and safety and consistency are top priorities
for us. If we ever really need the performance benefits of dropping down
to a lower isolation level, we can opt for that on a case-by-case basis
as the needs present themselves rather than endanger the whole
application with a weak default isolation level.