Split APK parsing out from APK set parsing
This change significantly reduces the complexity of the APK set parsing
method, making it easier to understand and review.

arrow-kt is used to simplify error handling. While using libraries that
change programming language paradigms is not something we typically do,
Kotlin is in serious need of more structured error handling, so we make
an exception for arrow-core here.