Fix incorrect ACL being checked in submitAppDraft()
When submitAppDraft() assigns a reviewer to an app draft, it must give
the assigned reviewer review access to the app draft. It does this by
creating an ACL with the canReview permission. However, there are some
cases where an ACL may already exist: namely, when the reviewer happens
to be the submitter (we will later prevent this from happening except in
specific cases such as for the Accrescent development team). In these
cases, we instead need to modify the existing ACL.

Unfortunately, submitAppDraft() modifies the existing ACL for the
_authenticated user_ instead of the _assigned reviewer_, resulting in a
privilege escalation vulnerability wherein a developer with submit
permissions for an app draft can bypass review by simply submitting the
app draft and subsequently creating an approving review for it with
their new review permission.

This isn't a particularly serious issue right now since this WIP version
of Parcelo isn't used in production and this bug would have almost
certainly been caught in final testing, but I thought it would make
sense to thoroughly document this issue so it doesn't come up again.