Remove named arguments for `PermissionService` calls
We originally used named arguments for all
`PermissionService.userCan*()` calls in the interest of ensuring
parameters were always passed in the right order to prevent
authorization bugs. However, the consequence of such a bug is almost
certainly not a false positive permission (privilege escalation) because
of our UUIDs' cryptographically random distribution, but instead an
easily identifiable false negative (permission denial).

Since the probability and impact of rolling out such a bug are both very
low and the complexity of the code is higher when using named arguments,
stop using named arguments for permission checks.