Create test users in Keycloak realm file
The default realm in Dev Services for Keycloak has a couple of problems
which make it unsuitable for our testing needs:

- It includes only a few users, limiting our options for testing
  user-specific resource constraints, permissions, etc.
- User "sub" fields are randomly generated, preventing us from having
  predictable identifiers to operate on

Instead of relying on the default realm users, create users from a
Keycloak realm file in development mode. This allows us to add and
remove testing users as needed and give them stable identifiers, solving
both of the above problems.